requirements.txt
# Interactive Brokers async API
ib_async>=1.0.1,<2

# Core data stack
pandas>=2.2,<3
duckdb>=1.1,<2
numpy>=1.26,<3

# Config / files
PyYAML>=6,<7
openpyxl>=3.1,<4
python-dateutil>=2.8,<3

# HTTP / utilities
requests>=2.31,<3

# Notebook / development
jupyter>=1,<2
ipykernel>=6,<7
notebook>=7,<8

# Visualization
matplotlib>=3.8,<4
